+++ modules/module-mssql/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @powersync/service-module-mssql
 
+## 0.8.1
+
+### Patch Changes
+
+- Updated dependencies [7e65360]
+  - @powersync/service-core@1.23.1
+
 ## 0.8.0
 
 ### Minor Changes